Define NOMINMAX to avoid annoying min() max() macros Create "cleangdiplus.h" which temporarily defines min/max, includes gdiplus.h, and undefs them Create HippoStdAfx.h for the common bits we always have in stdafx.h Link to STL so operator new will now throw bad_alloc Make lots of things throw bad_alloc rather than return HRESULT Add HippoMusicMonitor::getPrimingData(), not used/implemented yet Trap all exceptions from glib main loop dispatch handler (this is probably not useful, we need to trap our timeout/idle dispatch, not this) Add class HResultException so HRESULT can be thrown instead of returned Modify HippoBSTR to throw instead of returning HRESULTs, and redo setUTF8 Disable some pointless impossible-to-get-rid-off warnings with pragmas

+++ dumbhippo/trunk/client/windows/HippoUI/HippoFlickr.cpp
@@ -12,7 +12,7 @@
 extern "C" {
 #include <md5.h>
 }
-#include <gdiplus.h>
+#include "cleangdiplus.h"
 
 #import <msxml3.dll>  named_guids
 #include <mshtml.h>
@@ -271,7 +271,7 @@ HippoFlickr::computeAPISig(HippoArray<HippoBSTR> &paramNames, HippoArray<HippoBS
     utf = (unsigned char *) g_utf16_to_utf8(sig, -1, NULL, NULL, NULL);
     MD5Context md5Ctx;
     MD5Init(&md5Ctx);
-    MD5Update(&md5Ctx, utf, strlen((char*)utf));
+    MD5Update(&md5Ctx, utf, static_cast<unsigned int>(strlen((char*)utf)));
     MD5Final(digest, &md5Ctx);
     g_free(utf);
     for (unsigned int i = 0; i < 16; i++) {
